small fixes

+++ TankoholicClient/Bullet.cs
@@ -12,8 +12,8 @@ namespace TankoholicClient
         public Vector2 Position { get; private set; }
         private Vector2 velocity;
         private const int MULTIPLIER = 10;
-        public int Damage { get; set; }
-        public Bullet(Vector2 position, Vector2 direction, int damage = 5)
+        public float Damage { get; set; }
+        public Bullet(Vector2 position, Vector2 direction, float damage = 5)
         {
             Position = position;
             velocity = direction * MULTIPLIER;
